moved NotificationManager to Core and moved the create notification from Notification to NotificationManager
